+++ app/types/BookSchema.ts
@@ -1,0 +1,7 @@
+export interface BookSchema {
+    _id: { $oid: string };
+    title: string;
+    description: string;
+    author: string;
+    link: string;
+}