+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 # v.1.0.0
 
 - Add the core.
-- Add book endpoints.
+- Add book endpoints with mongoDB.